+++ sdks/python/geolens/models/__init__.py
@@ -451,7 +451,6 @@ from .reupload_preview_response_sample_rows_item import (
 )
 from .reupload_response import ReuploadResponse
 from .reupload_service_preview_request import ReuploadServicePreviewRequest
-from .saml_to_local_conversion import SamlToLocalConversion
 from .saved_search_create import SavedSearchCreate
 from .saved_search_create_params import SavedSearchCreateParams
 from .saved_search_list_response import SavedSearchListResponse
@@ -929,7 +928,6 @@ __all__ = (
     "ReuploadPreviewResponseSampleRowsItem",
     "ReuploadResponse",
     "ReuploadServicePreviewRequest",
-    "SamlToLocalConversion",
     "SavedSearchCreate",
     "SavedSearchCreateParams",
     "SavedSearchListResponse",